fixing folder structure
